Implement access request checking
To perform access request check, clients are required to submit full
specification of an action and resource that's being accessed. If the
provided key is present in the service storage, it is being decoded and
its scopes are analyzed to determine whether or not the request access
should be granted.

Signed-off-by: Dejan Mijic <dejan.mjc@gmail.com>